requirements.txt
flask==2.3.3
numpy==1.24.3
torch==2.0.1
opencv-python-headless==4.8.0.76
pydicom==2.4.3
Pillow==10.0.0
monai==1.2.0
torchvision==0.15.2
requests==2.31.0 